legg formkrav inn i egen mappe & små fiks av tilstands-feil
