+++ src/pages/klage/Klage.tsx
@@ -12,7 +12,7 @@ import { getPartialFramdriftsindikatorLinjeInfo } from '~utils/klage/klageUtils'
 import messages from './klage-nb';
 import styles from './klage.module.less';
 import SendKlageTilAttestering from './sendKlageTilAttestering/SendKlageTilAttestering';
-import VurderFormkrav from './VurderFormkrav';
+import VurderFormkrav from './vurderFormkrav/VurderFormkrav';
 import VurderingAvKlage from './vurderingAvKlage/VurderingAvKlage';
 
 const Klage = (props: { sak: Sak }) => {
