+++ src/pages/klage/klage.module.less
@@ -18,18 +18,12 @@
     width: 100%;
 }
 
-.container {
-    display: flex;
-    margin: @spacing-s 0;
-}
-
-.form {
+//sharedStyles
+.feilTilstandContainer {
     display: flex;
     flex-direction: column;
-    gap: @spacing-s;
-}
-
-.buttons {
-    display: flex;
-    gap: @spacing;
+    align-items: center;
+    gap: @spacing-xs;
+    margin-top: @spacing;
+    width: 80%;
 }
